pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel>=0.41.2"]
build-backend = "setuptools.build_meta"

[project]
name = "intentkit-py"
version = "0.4.0"
description = "An open-source Python library for building intent classification and execution systems that work with any AI backend."
authors = [
    {name = "Stephen Collins", email = "stephen@stephencollins.tech"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://github.com/Stephen-Collins-tech/intent-kit"
Documentation = "https://docs.intentkit.io"
Repository = "https://github.com/Stephen-Collins-tech/intent-kit"
"Bug Tracker" = "https://github.com/Stephen-Collins-tech/intent-kit/issues"

[project.optional-dependencies]
all = [
    "anthropic>=0.54.0",
    "google-genai>=0.1.0",
    "ollama>=0.1.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.2",
]
anthropic = [
    "anthropic>=0.54.0",
]
google = [
    "google-genai>=0.1.0",
]
ollama = [
    "ollama>=0.1.0",
]
openai = [
    "openai>=1.0.0",
]
yaml = [
    "pyyaml>=6.0.2",
]

[project.scripts]
eval-node = "intent_kit.evals.run_node_eval:main"
run-evals = "intent_kit.evals.run_all_evals:run_all_evaluations"
lint = "scripts.lint:main"
typecheck = "scripts.typecheck:main"
examples = "scripts.examples:run_all"
example = "scripts.examples:run_single"
list-examples = "scripts.examples:list_examples"

[tool.setuptools.packages.find]
where = ["."]
include = ["intent_kit*", "scripts*"]

[tool.uv.sources]
intentkit-py = { workspace = true }

[dependency-groups]
dev = [
    "black[jupyter]>=24.4.2",
    "black>=24.4.2",
    "build>=1.2.1",
    "coverage>=7.0",
    "intentkit-py[all]",
    "ipykernel>=6.0.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-material>=9.5.17",
    "mkdocs>=1.5.0",
    "mkdocstrings[python]>=0.24.0",
    "mypy>=1.10.0",
    "pre-commit>=3.6.0",
    "pytest-cov>=5.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "pyyaml",
    "ruff>=0.4.7",
    "tqdm",
    "twine>=5.1.0",
    "types-pyyaml>=6.0.12.20250516",
]
